+++ CHANGELOG.md
@@ -3,4 +3,5 @@
 ## 0.1.13
 
 - added image tooltip, showing x, y and channel data.
+- added selection box on click.
 - refactored display widgets